src/py_requirements.txt
altgraph==0.16.1
attrs==19.3.0
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==20.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
PyInstaller==3.6
pyparsing==2.4.6
PySide2==5.14.0
pytest==5.3.2
pyusb==1.0.2
pyzmq==18.1.1
shiboken2==5.14.0
six==1.13.0
wcwidth==0.1.8
zipp==0.6.0
